Fixes make sources
In the Makefile we specify the directories that will be included in
the tarball. This means that everytime we add a new directory we
have to update the Makefile.

Furthermore, the tarball does not include .git, which breaks the
building process when we attempt to build app.

With this fix we add all the files to the tarball.